Fixes #26109 - Add "Web Console" button and support
This adds a "Web Console" button next to the "Schedule Remote Job"
button in the title row of a host page.

The button links to a specially configured Cockpit instance.  It's
location is configured with the new remote_execution_cockpit_href
setting.

The Cockpit instance will use the new /cockpit_redirect path to get a
token via OAuth style redirection.